+++ src/assets/styles/StepsForm.scss
@@ -1,4 +1,4 @@
-@import '../Variables.scss'; 
+@import '../styles/Variables.scss'; 
 
 // principal 
 .container-Foodbank{  
@@ -9,6 +9,7 @@
     overflow: hidden;
     position: absolute;
     display: inline-block;
+    //margin-top: 10%;
 }
 
 // estilos imagen principal 
@@ -20,6 +21,44 @@
     position: absolute;
 }
 
+.containerText {
+    //margin: 50% 23% 5% 23%;
+    width: 48vw;
+    height: 94px;
+    //left: 40.41vw;
+    font-family: Montserrat;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 56px;
+    line-height: 80px;
+    color: #FFFFFF;
+    position: absolute;
+    margin-top: 30%; 
+    left: 50%;
+    transform: translate(-50%, -50%);      
+}
+
+.containerText2 {
+    //margin: 50% 23% 5% 23%;
+    font-family: Montserrat;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 36px;
+    line-height: 50px;
+    text-align: center;
+    letter-spacing: -0.015em;
+    color: #333333;
+    margin-bottom: 10vh;
+
+}
+// para las pasos se toma los mismos estilos de detail y se llama desde alli
+
+.containerFoodDetails{
+    width: 100vw;
+    height: 430px;
+    margin-top: 130vh;
+    position: absolute;
+}
 .Textsteps{
     width: 100vw;
     margin-bottom: 7vh;
@@ -32,7 +71,29 @@
     letter-spacing: -0.015em;
     color: #FF8150;    
 }
-// para las pasos se toma los mismos estilos de detail y se llama desde alli
+.containerfood1 {
+    margin-left: 2vw;
+    width: 7.36vw;
+    height: 7.36vw;
+    text-align: center;
+    font-family: Montserrat;
+    line-height: 2.76vw;
+    letter-spacing: -0.015em;
+    border-radius: 50%;
+    background: linear-gradient(180deg, #FCC3AD 17.76%, rgba(239, 203, 190, 0) 93.8%);
+}
+
+.imgfoodDetail {
+    width: 40%;
+    height: 40%;
+    margin-top: 3.35rem;
+    margin-bottom: 2vw;
+}
+
+.formulario{
+    color:$textColor5;
+}
+  
 
 // contenedor de imagenes
 .container-section-imgText{
@@ -157,7 +218,7 @@
     width: 100vw;
     height: 459px;
     margin-bottom: 0.1px;
-    background: url(../../static/images/imgContactForm.jpg);
+    //background: url(../../static/images/imgContactForm.jpg);
     background-repeat: no-repeat;
     background-size: cover;
