+++ src/assets/styles/components/ContacForm.scss
@@ -55,8 +55,8 @@
     display:flex;
     width: 100vw;
     height: 459px;
-
-    margin-bottom: 0.1px;
+    margin-top: 6749px;
+    margin-bottom: 500px;
     background: url(../../static/images/imgContactForm.jpg);
     background-repeat: no-repeat;
     background-size: cover;
