+++ src/assets/styles/components/FoodBankForm.scss
@@ -1,0 +1,93 @@
+@import '../Variables.scss'; 
+
+// box imput formulario 
+.textInputFood{
+    font-size: 1rem;
+    font-weight: 500;
+    font-weight: bold;
+    color:$textColor4;
+
+}
+.inputBoxFood {
+    width: 615px;
+    font-size: 0.80rem;
+    padding: 20px;
+    color: #FFFFFF;
+    margin-top: 10%;
+    border: 1px solid #BDBDBD;
+}
+.container-Food{
+    flex-direction: column;
+    //margin-left: 20vw;
+    //margin-top: 20vh;
+    position: absolute;
+    width: 672px;
+    height: 664px;
+    left: 212px;
+    top: 1946px;
+
+}
+// textos  seccion  formulario de contacto 
+.text1Food{
+    color: $textColor5 ;
+    font-family: Montserrat;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 24px;
+    line-height: 36px;
+    text-align: left;
+    letter-spacing: -0.015em;
+}
+
+.text2Food{
+    color: $textColor3;
+    font-size:1.5rem;
+    font-weight: 400;
+    position: absolute;
+    width: 417px;
+    height: 312px;
+    left: 1018px;
+    //top: 2114px;
+    font-family: Montserrat;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 18px;
+    line-height: 30px;
+}
+.container-foodtexts{
+    margin-top: 10vh;
+    margin-left: 5vw;
+    width: 670px;
+}
+// container  seccion 
+.container-seccionFood{
+    margin-bottom: 10%;
+    position: absolute;
+    width: 1520px;
+    height: 812px;
+    left: 80px;
+    //top: 1874px;
+    background: #FFFFFF;
+    //box-shadow: 0px 4px 20px rgba(0, 0, 0, 0.25)
+}
+// estilos  boton 
+
+.btnSend{
+    width: 185px;
+    height:50px ;
+    left: 672px;
+    color:$textColor5;
+    background-color: $textColor3;
+    font-size: 1.125rem;
+    text-align: center;
+    align-self: center;
+    overflow: hidden;
+    position: absolute;
+    //top: 2538px;
+    border: 1px solid #828282;
+    border-radius: 5px;
+}
+
+.container-btnSend{
+    margin-top: 5vh;
+}
