+++ src/components/Food Bank/ StepsFBank.jsx
@@ -6,12 +6,15 @@ import listaverificacion from "../../assets/static/images/imgFoodBank/IconFoodBa
 import tiempoentrega from "../../assets/static/images/imgFoodBank/IconFoodBank/tiempoentrega.png";
 import transportando from "../../assets/static/images/imgFoodBank/IconFoodBank/transportando.png";
 import '../../assets/styles/Detail.scss'
-import '../../assets/styles/components/StepsForm.scss'
+import '../../assets/styles/StepsForm.scss'
 
 function Steps() {
   return (
     <Fragment>
-      <div className="containerDetails">
+      <div className="containerFoodDetails">
+        <div className="containerText2">
+          <p>"Creemos que <strong>los alimentos no deben ser desperdicios,</strong> por ello los que te sobre en casa o no ocupes, <strong>aquí son bienvenidos."</strong></p>
+        </div>
         <div className="Textsteps">
           <p>Sigue estos 6 pasos, para poder colaborar.</p>
         </div>
@@ -19,8 +22,8 @@ function Steps() {
         <div className="containerBoxDetails">
 
           <div className="boxContainerDetail">
-            <div className="container1">
-              <img src={listaverificacion} className="imgDetail" alt={"Imagen"} />
+            <div className="containerfood1">
+              <img src={listaverificacion} className="imgfoodDetail" alt={"Imagen"} />
             </div>
             <ul>
               <p>Haz una lista de los alimentos con los que deseas colaborar, que sean no perecederos o congelados.</p>
@@ -28,8 +31,8 @@ function Steps() {
           </div>
 
           <div className="boxContainerDetail">
-            <div className="container1">
-              <img src={formulariocontacto} className="imgDetail" alt={"Imagen"} />
+            <div className="containerfood1">
+              <img src={formulariocontacto} className="imgfoodDetail" alt={"Imagen"} />
             </div>
             <ul>
               <p>Rellena con tus datos el formulario a continuación y envíalo a través del botón “colabora”.</p>
@@ -37,8 +40,8 @@ function Steps() {
           </div>
 
           <div className="boxContainerDetail">
-            <div className="container1">
-              <img src={tiempoentrega} className="imgDetail" alt={"Imagen"} />
+            <div className="containerfood1">
+              <img src={tiempoentrega} className="imgfoodDetail" alt={"Imagen"} />
             </div>
             <ul>
               <p>Nos comunicaremos contigo, por medio de correo electrónico o llamada telefónica, para coordinar la entrega de alimentos.</p>
@@ -46,8 +49,8 @@ function Steps() {
           </div>
 
           <div className="boxContainerDetail">
-            <div className="container1">
-              <img src={cajas} className="imgDetail" alt={"Imagen"} />
+            <div className="containerfood1">
+              <img src={cajas} className="imgfoodDetail" alt={"Imagen"} />
             </div>
             <ul>
               <p>En cajas guarda los alimentos con los que vas a colaborar, para facilitar su traslado.</p>
@@ -55,8 +58,8 @@ function Steps() {
           </div>
 
           <div className="boxContainerDetail">
-            <div className="container1">
-              <img src={transportando} className="imgDetail" alt={"Imagen"} />
+            <div className="containerfood1">
+              <img src={transportando} className="imgfoodDetail" alt={"Imagen"} />
             </div>
             <ul>
               <p>Trasládalos  hasta nuestras áreas de distribución, si no nosotros haremos el traslado de los alimentos.</p>
@@ -64,8 +67,8 @@ function Steps() {
           </div>
 
           <div className="boxContainerDetail">
-            <div className="container1">
-              <img src={familia} className="imgDetail" alt={"Imagen"} />
+            <div className="containerfood1">
+              <img src={familia} className="imgfoodDetail" alt={"Imagen"} />
             </div>
             <ul>
               <p>Finalmente, haremos llegar los alimentos a los restaurantes, quienes cocinarán ricos almuerzos a las comunidades vulnerables.</p>
