+++ src/components/Food Bank/FoodBankForm/FoodBanckFormText.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const ContacFormText = ({ text1Form, text2Form }) => {
+  
+  return (
+    <div className="container-foodtexts">
+      <div>
+        <p className="text2Food">{text2Form} </p>
+      </div>
+    </div>
+  );
+}
+
+export default ContacFormText;
